Move integration tests into beeflow package
